GigaChatAPI/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
exceptiongroup==1.3.0
Flask==3.1.2
gigachat==0.1.42.post2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.79
langchain-gigachat==0.3.12
langsmith==0.4.41
MarkupSafe==3.0.3
orjson==3.11.4
packaging==25.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.2
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zstandard==0.25.0
